Energy net rework

+++ src/ic2/common/BlockCable.java
@@ -34,16 +34,16 @@ public class BlockCable extends BlockMultiID implements IPaintableBlock, ISpecia
     Ic2Items.detectorCableBlock = new ItemStack(this, 1, 11);
     Ic2Items.splitterCableBlock = new ItemStack(this, 1, 12);
   }
-
+  
   public static int getCableColor(IBlockAccess iblockaccess, int i, int j, int k) {
     TileEntity tileentity = iblockaccess.getTileEntity(i, j, k);
     return tileentity instanceof TileEntityCable ? ((TileEntityCable) tileentity).color : 0;
   }
-
+  
   public String getTextureFile() {
     return "/ic2/sprites/block_cable.png";
   }
-
+  
   public int getBlockTexture(IBlockAccess iblockaccess, int i, int j, int k, int l) {
     short word1 = 0;
     TileEntity tileentity = iblockaccess.getTileEntity(i, j, k);
@@ -54,10 +54,10 @@ public class BlockCable extends BlockMultiID implements IPaintableBlock, ISpecia
         if (tileentitycable.foamed == 1) {
           return 178;
         }
-
+  
         return 208 + tileentitycable.foamColor;
       }
-
+  
       word0 = tileentitycable.cableType;
       if (!(tileentity instanceof TileEntityCableDetector) && !(tileentity instanceof TileEntityCableSplitter)) {
         word1 = tileentitycable.color;
@@ -69,14 +69,14 @@ public class BlockCable extends BlockMultiID implements IPaintableBlock, ISpecia
     else {
       word0 = (short) iblockaccess.getData(i, j, k);
     }
-
+    
     return word0 * 16 + word1;
   }
-
+  
   public int a(int i, int j) {
     return j * 16;
   }
-
+  
   public MovingObjectPosition a(World world, int i, int j, int k, Vec3D vec3d, Vec3D vec3d1) {
     TileEntity tileentity = world.getTileEntity(i, j, k);
     if (!(tileentity instanceof TileEntityCable)) {
@@ -89,67 +89,82 @@ public class BlockCable extends BlockMultiID implements IPaintableBlock, ISpecia
       double d1 = (double) tileentitycable.getCableThickness() / 2.0D;
       boolean flag = false;
       Vec3D vec3d3 = Vec3D.a(0.0D, 0.0D, 0.0D);
-      Direction direction = AabbUtil.getIntersection(vec3d, vec3d2, AxisAlignedBB.a((double) i + 0.5D - d1, (double) j + 0.5D - d1, (double) k + 0.5D - d1, (double) i + 0.5D + d1, (double) j + 0.5D + d1, (double) k + 0.5D + d1), vec3d3);
+      Direction direction = AabbUtil.getIntersection(vec3d, vec3d2, AxisAlignedBB
+          .a((double) i + 0.5D - d1, (double) j + 0.5D - d1, (double) k + 0.5D - d1, (double) i + 0.5D + d1,
+              (double) j + 0.5D + d1, (double) k + 0.5D + d1), vec3d3);
       if (direction != null && vec3d3.b(vec3d) <= d) {
         flag = true;
       }
-
+  
       Direction[] adirection = Direction.values();
       int l = adirection.length;
-
+  
       for (int i1 = 0; i1 < l; ++i1) {
         Direction direction1 = adirection[i1];
         if (flag) {
           break;
         }
-
+    
         TileEntity tileentity1 = direction1.applyToTileEntity(tileentitycable);
         if (tileentitycable.canInteractWith(tileentity1)) {
           AxisAlignedBB axisalignedbb = null;
           switch (direction1.ordinal() + 1) {
             case 1:
-              axisalignedbb = AxisAlignedBB.a((double) i, (double) j + 0.5D - d1, (double) k + 0.5D - d1, (double) i + 0.5D, (double) j + 0.5D + d1, (double) k + 0.5D + d1);
+              axisalignedbb = AxisAlignedBB
+                  .a(i, (double) j + 0.5D - d1, (double) k + 0.5D - d1, (double) i + 0.5D,
+                      (double) j + 0.5D + d1, (double) k + 0.5D + d1);
               break;
             case 2:
-              axisalignedbb = AxisAlignedBB.a((double) i + 0.5D, (double) j + 0.5D - d1, (double) k + 0.5D - d1, (double) i + 1.0D, (double) j + 0.5D + d1, (double) k + 0.5D + d1);
+              axisalignedbb = AxisAlignedBB
+                  .a((double) i + 0.5D, (double) j + 0.5D - d1, (double) k + 0.5D - d1, (double) i + 1.0D,
+                      (double) j + 0.5D + d1, (double) k + 0.5D + d1);
               break;
             case 3:
-              axisalignedbb = AxisAlignedBB.a((double) i + 0.5D - d1, (double) j, (double) k + 0.5D - d1, (double) i + 0.5D + d1, (double) j + 0.5D, (double) k + 0.5D + d1);
+              axisalignedbb = AxisAlignedBB
+                  .a((double) i + 0.5D - d1, j, (double) k + 0.5D - d1, (double) i + 0.5D + d1,
+                      (double) j + 0.5D, (double) k + 0.5D + d1);
               break;
             case 4:
-              axisalignedbb = AxisAlignedBB.a((double) i + 0.5D - d1, (double) j + 0.5D, (double) k + 0.5D - d1, (double) i + 0.5D + d1, (double) j + 1.0D, (double) k + 0.5D + d1);
+              axisalignedbb = AxisAlignedBB
+                  .a((double) i + 0.5D - d1, (double) j + 0.5D, (double) k + 0.5D - d1, (double) i + 0.5D + d1,
+                      (double) j + 1.0D, (double) k + 0.5D + d1);
               break;
             case 5:
-              axisalignedbb = AxisAlignedBB.a((double) i + 0.5D - d1, (double) j + 0.5D - d1, (double) k, (double) i + 0.5D + d1, (double) j + 0.5D, (double) k + 0.5D);
+              axisalignedbb = AxisAlignedBB
+                  .a((double) i + 0.5D - d1, (double) j + 0.5D - d1, k, (double) i + 0.5D + d1,
+                      (double) j + 0.5D, (double) k + 0.5D);
               break;
             case 6:
-              axisalignedbb = AxisAlignedBB.a((double) i + 0.5D - d1, (double) j + 0.5D - d1, (double) k + 0.5D, (double) i + 0.5D + d1, (double) j + 0.5D + d1, (double) k + 1.0D);
+              axisalignedbb = AxisAlignedBB
+                  .a((double) i + 0.5D - d1, (double) j + 0.5D - d1, (double) k + 0.5D, (double) i + 0.5D + d1,
+                      (double) j + 0.5D + d1, (double) k + 1.0D);
           }
-
+  
           direction = AabbUtil.getIntersection(vec3d, vec3d2, axisalignedbb, vec3d3);
           if (direction != null && vec3d3.b(vec3d) <= d) {
             flag = true;
           }
         }
       }
-
+  
       return flag ? new MovingObjectPosition(i, j, k, direction.toSideValue(), vec3d3) : null;
     }
   }
-
+  
   public AxisAlignedBB getCollisionBoundingBoxFromPool(World world, int i, int j, int k, int l) {
-    double d = (double) TileEntityCable.getCableThickness(l);
-    return AxisAlignedBB.b((double) i + 0.5D - d, (double) j + 0.5D - d, (double) k + 0.5D - d, (double) i + 0.5D + d, (double) j + 0.5D + d, (double) k + 0.5D + d);
+    double d = TileEntityCable.getCableThickness(l);
+    return AxisAlignedBB.b((double) i + 0.5D - d, (double) j + 0.5D - d, (double) k + 0.5D - d, (double) i + 0.5D + d,
+        (double) j + 0.5D + d, (double) k + 0.5D + d);
   }
-
+  
   public AxisAlignedBB e(World world, int i, int j, int k) {
     return this.getCommonBoundingBoxFromPool(world, i, j, k, false);
   }
-
+  
   public AxisAlignedBB getSelectedBoundingBoxFromPool(World world, int i, int j, int k) {
     return this.getCommonBoundingBoxFromPool(world, i, j, k, true);
   }
-
+  
   public AxisAlignedBB getCommonBoundingBoxFromPool(World world, int i, int j, int k, boolean flag) {
     TileEntity tileentity = world.getTileEntity(i, j, k);
     if (!(tileentity instanceof TileEntityCable)) {
@@ -165,50 +180,50 @@ public class BlockCable extends BlockMultiID implements IPaintableBlock, ISpecia
       double d5 = (double) j + 0.5D + d;
       double d6 = (double) k + 0.5D + d;
       if (tileentitycable.canInteractWith(world.getTileEntity(i - 1, j, k))) {
-        d1 = (double) i;
+        d1 = i;
       }
-
+  
       if (tileentitycable.canInteractWith(world.getTileEntity(i, j - 1, k))) {
-        d2 = (double) j;
+        d2 = j;
       }
-
+  
       if (tileentitycable.canInteractWith(world.getTileEntity(i, j, k - 1))) {
-        d3 = (double) k;
+        d3 = k;
       }
-
+  
       if (tileentitycable.canInteractWith(world.getTileEntity(i + 1, j, k))) {
-        d4 = (double) (i + 1);
+        d4 = i + 1;
       }
-
+  
       if (tileentitycable.canInteractWith(world.getTileEntity(i, j + 1, k))) {
-        d5 = (double) (j + 1);
+        d5 = j + 1;
       }
-
+  
       if (tileentitycable.canInteractWith(world.getTileEntity(i, j, k + 1))) {
-        d6 = (double) (k + 1);
+        d6 = k + 1;
       }
-
+  
       return AxisAlignedBB.b(d1, d2, d3, d4, d5, d6);
     }
   }
-
+  
   public boolean isBlockNormalCube(World world, int i, int j, int k) {
     TileEntity tileentity = world.getTileEntity(i, j, k);
     if (tileentity instanceof TileEntityCable) {
       TileEntityCable tileentitycable = (TileEntityCable) tileentity;
       return tileentitycable.foamed > 0;
     }
-
+    
     return false;
   }
-
+  
   public boolean interact(World world, int i, int j, int k, EntityHuman entityhuman) {
     ItemStack itemstack = entityhuman.U();
     if (itemstack != null && itemstack.id == Block.SAND.id) {
       if (!Platform.isSimulating()) {
         return true;
       }
-
+  
       TileEntity tileentity = world.getTileEntity(i, j, k);
       if (tileentity instanceof TileEntityCable) {
         TileEntityCable tileentitycable = (TileEntityCable) tileentity;
@@ -217,23 +232,23 @@ public class BlockCable extends BlockMultiID implements IPaintableBlock, ISpecia
           if (itemstack.count <= 0) {
             entityhuman.inventory.items[entityhuman.inventory.itemInHandIndex] = null;
           }
-
+  
           return true;
         }
       }
     }
-
+    
     return false;
   }
-
+  
   public boolean colorBlock(World world, int i, int j, int k, int l) {
     return ((TileEntityCable) world.getTileEntity(i, j, k)).changeColor(l);
   }
-
+  
   public boolean canHarvestBlock(EntityHuman entityhuman, int i) {
     return true;
   }
-
+  
   public ArrayList getBlockDropped(World world, int i, int j, int k, int l, int i1) {
     ArrayList arraylist = new ArrayList();
     TileEntity tileentity = world.getTileEntity(i, j, k);
@@ -244,21 +259,22 @@ public class BlockCable extends BlockMultiID implements IPaintableBlock, ISpecia
     else if (l != 13) {
       arraylist.add(new ItemStack(Ic2Items.insulatedCopperCableItem.id, 1, l));
     }
-
+    
     return arraylist;
   }
-
+  
   public void remove(World world, int i, int j, int k) {
     if (world.getData(i, j, k) == 13) {
       TileEntity tileentity = world.getTileEntity(i, j, k);
       if (tileentity instanceof TileEntityCable) {
         TileEntityCable tileentitycable = (TileEntityCable) tileentity;
-        StackUtil.dropAsEntity(world, i, j, k, new ItemStack(Ic2Items.insulatedCopperCableItem.id, 1, tileentitycable.cableType));
+        StackUtil.dropAsEntity(world, i, j, k,
+            new ItemStack(Ic2Items.insulatedCopperCableItem.id, 1, tileentitycable.cableType));
       }
     }
-
+    
   }
-
+  
   public TileEntityBlock getBlockEntity(int i) {
     if (i == 11) {
       return new TileEntityCableDetector((short) i);
@@ -267,23 +283,23 @@ public class BlockCable extends BlockMultiID implements IPaintableBlock, ISpecia
       return i == 12 ? new TileEntityCableSplitter((short) i) : new TileEntityCable((short) i);
     }
   }
-
+  
   public int getMaxMeta() {
     return 13;
   }
-
+  
   public boolean b() {
     return false;
   }
-
+  
   public int c() {
     return mod_IC2.cableRenderId;
   }
-
+  
   public boolean a() {
     return false;
   }
-
+  
   public void attack(World world, int i, int j, int k, EntityHuman entityhuman) {
     if (entityhuman.U() != null && entityhuman.U().getItem() instanceof ItemToolCutter) {
       Entity bukkitentity = entityhuman.getBukkitEntity();
@@ -294,28 +310,29 @@ public class BlockCable extends BlockMultiID implements IPaintableBlock, ISpecia
         if (breakev.isCancelled()) {
           return;
         }
-
+  
         breakev.setCancelled(true);
       }
-
+  
       ItemToolCutter.cutInsulationFrom(entityhuman.U(), world, i, j, k);
     }
-
+    
   }
-
+  
   public boolean a(IBlockAccess iblockaccess, int i, int j, int k, int l) {
     TileEntity tileentity = iblockaccess.getTileEntity(i, j, k);
     return tileentity instanceof TileEntityCableDetector && ((TileEntityCableDetector) tileentity).getActive();
   }
-
+  
   public void addCreativeItems(ArrayList arraylist) {
   }
-
+  
   public float getHardness(int i) {
     return i != 13 ? 0.2F : 3.0F;
   }
-
-  public float getSpecialExplosionResistance(World world, int i, int j, int k, double d, double d1, double d2, net.minecraft.server.Entity entity) {
+  
+  public float getSpecialExplosionResistance(World world, int i, int j, int k, double d, double d1, double d2,
+                                             net.minecraft.server.Entity entity) {
     TileEntity tileentity = world.getTileEntity(i, j, k);
     if (tileentity instanceof TileEntityCable) {
       TileEntityCable tileentitycable = (TileEntityCable) tileentity;
@@ -323,10 +340,10 @@ public class BlockCable extends BlockMultiID implements IPaintableBlock, ISpecia
         return 90.0F;
       }
     }
-
+    
     return 6.0F;
   }
-
+  
   public boolean canConnectRedstone(IBlockAccess iblockaccess, int i, int j, int k, int l) {
     int i1 = iblockaccess.getData(i, j, k);
     return i1 == 11 || i1 == 12;
