+++ src/ic2/common/BlockMiningPipe.java
@@ -12,32 +12,34 @@ public class BlockMiningPipe extends BlockTex {
     ModLoader.registerBlock(this, ItemBlockCommon.class);
     Ic2Items.miningPipe = new ItemStack(this);
   }
-
+  
   public boolean canPlace(World world, int i, int j, int k) {
     return false;
   }
-
+  
   public boolean a() {
     return false;
   }
-
+  
   public boolean isBlockNormalCube(World world, int i, int j, int k) {
     return false;
   }
-
+  
   public boolean b() {
     return false;
   }
-
+  
   public int c() {
     return mod_IC2.miningPipeRenderId;
   }
-
+  
   public AxisAlignedBB e(World world, int i, int j, int k) {
-    return AxisAlignedBB.b((double) ((float) i + 0.375F), (double) ((float) j), (double) ((float) k + 0.375F), (double) ((float) i + 0.625F), (double) ((float) j + 1.0F), (double) ((float) k + 0.625F));
+    return AxisAlignedBB.b((float) i + 0.375F, (float) j, (float) k + 0.375F,
+        (float) i + 0.625F, (float) j + 1.0F, (float) k + 0.625F);
   }
-
+  
   public AxisAlignedBB getSelectedBoundingBoxFromPool(World world, int i, int j, int k) {
-    return AxisAlignedBB.b((double) ((float) i + 0.375F), (double) ((float) j), (double) ((float) k + 0.375F), (double) ((float) i + 0.625F), (double) ((float) j + 1.0F), (double) ((float) k + 0.625F));
+    return AxisAlignedBB.b((float) i + 0.375F, (float) j, (float) k + 0.375F,
+        (float) i + 0.625F, (float) j + 1.0F, (float) k + 0.625F);
   }
 }
